clip/requirements.txt
fastapi
Pillow
python-multipart
transformers
uvicorn[standard]
numpy
nmslib